Created index.html for first project
Created an Index.html file which includes the CSS code and JavaScript code attribution for the code used in the project is included in the index.html file prior to the start of the actual web code. I used w3schools.org for the 2nd to final section which was the "About Me" modal popup box the box when clicked shows a Modal that includes information about my goals and what my main intention is with enrolling with Intro to HTML 5 course.The rest of the code was used with accessibility.Unfortunately the Modal boxes had to be made using Div containers but will be updated if needed to reflect accessibility standards. Thank you.